preliminary class refactoring of neon_site into base_site and tower_site
